ci: add github ci workflow
Adds a GitHub CI workflow that contains stages to lint, test and build
the charm. These stages assume that the runner is an Ubuntu 22.04
container or VM and that the charmcraft and uv packages are available.
Currently the runner environment is set up as an Incus image locally
available in the runner machine, which uses GARM to dynamically spawn a
new container for each job.

The charmcraft package is needed to build the charm, and is installed in
the runner as a snap. The uv package is needed to install python tools
such as tox, which is used as an entrypoint to test charms. Tools
installed with uv can be executed with the uvx command.

References:
https://github.com/cloudbase/garm
https://docs.astral.sh/uv/concepts/tools/
Signed-off-by: Luís Simas <luissimas@protonmail.com>