ci: add integration test job
Adds an integration test job on the CI workflow. This job bootstraps a
local Juju controller and LXD cloud to run the integration tests. The
tests use the charm file built in the build job to perform the tests.
This ensures that we test against the actual charm artifact that will be
exposed, while also avoiding building the charm again in the integration
test job.

Signed-off-by: Luís Simas <luissimas@protonmail.com>