feat: basic cluster formation
Enables a basic Incus cluster formation when multiple units are
deployed.

A new peer relation called cluster is introduced to share node names and
tokens between units. The leader unit is the bootstrap server used to
generate the tokens for the remaining units. The remaining units get a
token from the peer relation data and use it to join the cluster.

Signed-off-by: Luís Simas <luissimas@protonmail.com>