fix: check for unit data only on leader
Checks if the unit data exists only on the leader unit when handling the
cluster-relation-changed event. The reason for this is that non-leader
units might get this event in response to a change in the application
databag, which is the case when the leader unit puts a new token in the
application data. When this is the case, no unit is associated with the
event and therefore the unit data associated with the event does not
exist.

Signed-off-by: Luís Simas <luissimas@protonmail.com>