feat: add ceph relation
Adds a ceph relation via the ceph-client interface to the charm. This
relation is used to provide a Ceph storage pool to the Incus cluster.

The ceph-common package is installed on the unit if a ceph relation is
present. The installation can occur in the install event if the relation
is already present at that stage. If the relation with ceph is added
only after the deployment, the charm checks the support for Ceph in the
current Incus instance and install the packages if needed.

The relation data is used to write the ceph config files. These files
are then used by Incus to communicate with the Ceph cluster.

Signed-off-by: Luís Simas <luissimas@protonmail.com>