fix: check server certificate before joining cluster
Units now check the current server certificate before joining the
cluster. If the certificate is not yet applied to the server, the event
is deferred. This prevents a race condition in which the units would try
to join the cluster with a self-signed certificate while the bootstrap
server expected a certificate signed by an external CA such as vault.

Signed-off-by: Luís Simas <luissimas@protonmail.com>