feat: create ceph storage pool
Now the Ceph storage pool is created automatically by the charm via the
ceph-client interface. When this creation is orchestrated in a clustered
environment to ensure that we create the pending storage pool on all
Incus nodes and also avoid race conditions with new members joining
while the pool is being created.

Signed-off-by: Luís Simas <luissimas@protonmail.com>