feat: set cluster node failure domains based on juju availability zone
Adds support for setting cluster member failure domains based on Juju's
availability zone. This allows operators to use the availability zone of
the underlying Juju cloud as the source of truth for incus failure
domains.

The behavior is enabled by default, but a config option
set-failure-domain is also provided as a way to opt-out of this
behavior. If the config option is set to true and the environment
variable JUJU_AVAILABILITY_ZONE is set on the unit, its value is used to
set the Incus cluster member failure domain property.

Signed-off-by: Luís Simas <luissimas@protonmail.com>