fix: set ovn connection options only when all requirements are present
Now any Incus config option related to the OVN connection is set only if
all requirements for the connection are present. This is needed since
Incus performs a connection check after changing any of those values, so
we must ensure that we only set those if we have everything needed for a
working connection.

The requirements for a working connection with OVN are:

- Client SSL certificates issued by vault
- Northbound endpoints received from all ovn-central units

If any of the requirements are not met, the charm does not set any OVN
related config option in the current event.

Signed-off-by: Luís Simas <luissimas@protonmail.com>