ci: use separate runner pool for integration tests
Adds tags to the checks workflow to allow running integration tests and
lightweight checks in separate pools. This allows us to limit the number
of concurrent integration tests to 1, while still allowing many lightweight
checks to be ran concurrently.

This is needed since integration tests need up to 16GB of RAM on the
runner containers, and we have limited resources on our self-hosted
runners.

Signed-off-by: Luís Simas <luissimas@protonmail.com>