fix: set cluster relation data on joined event
We've observed that in some environments Juju does not emit the
relation-created event reliably. This caused the leader unit to never
setup the application data correctly, leading to a deadlock when forming
the cluster. By moving the setting of application data to the
relation-joined event, we hope to be more resilient to missing events.

Signed-off-by: Luís Simas <luissimas@protonmail.com>