feat: automatic ovn network creation
Now the charm is able to automatically create an OVN network and its
uplink on the Incus cluster.

This implementation accounts for three different cases of OVN network
creation:

1. The ovsdb-cms relation is established after the Incus cluster is formed
2. The ovsdb-cms relation is established before the Incus cluster is
   formed
3. A new node is added into an existing Incus cluster in which a OVN
   network was already created by the charm

The first two cases are treated as a single case in the implementation.
In such cases, the creation of the OVN network consists in first
creating the uplink network on all cluster members, then instantiating
this network in the cluster and finally creating the OVN network
associating it with the uplink.

In the third case, the joining members include the uplink network in
their bootstrap data to ensure that all cluster members have an uplink
network setup.

Signed-off-by: Luís Simas <luissimas@protonmail.com>