fix: set cms-client-bound-address on relation joined event
We've observed that in some environments Juju does not emit the
relation-created event reliably. This caused the unit to never set its
IP address on the relation data, which means that ovn-central never
created the appropriate firewall rules for the Incus unit, ultimately
leading to a failure in the OVN northbound connection configuration.
By moving the setting of application data to the relation-joined event,
we hope to be more resilient to missing events.

Signed-off-by: Luís Simas <luissimas@protonmail.com>