fix: check for unit in data bag on certificates relation
Due to a recent change in the way Juju handles relation data, when
dealing with a relation departed event, it is possible that the
operation of getting the relation data for a unit within a given
relation might fail. To handle this, we now check if the data is
available and skip the iteration if it isn't.

Signed-off-by: Luís Simas <luissimas@protonmail.com>