test: assert that failure domain is other than the default on all members
When we first wrote these tests, the LXD cloud (which we use in our CI)
did not support availability zones. For this reason, we asserted that
all nodes had the "default" failure domain.

Juju 3.6.9 introduced support for availability zones in LXD clouds and
sets them to the host name that is running the LXD VMs. With that in
mind, we need to update the integration test to instead assert that we
have a failure domain different than the default. Since our CI creates
LXD nodes on-demand, it makes little sense to match an specific
hostname.

References:
https://documentation.ubuntu.com/juju/3.6/reference/juju/juju-roadmap-and-releases/#lxd

Signed-off-by: Luís Simas <luissimas@protonmail.com>