ci: explicitly install required tools
Explicitly install all required CI dependencies on the runners. Before
this, we were using a custom image based on Ubuntu Jammy that had some
packages pre-installed. By making everything explicitly installed in the
workflow file, and depending only on a base Ubuntu Noble image, we
increase the portability of the CI for other environments.

Signed-off-by: Luís Simas <luissimas@protonmail.com>